src/main/java/com/btcplatform/trading/service/UserService.java
package com.btcplatform.trading.service;

import com.btcplatform.trading.model.User;
import com.btcplatform.trading.repository.UserRepository;
import org.springframework.beans.factory.annotation.Autowired;
import org.springframework.stereotype.Service;

import java.math.BigDecimal;

@Service
public class UserService {

    @Autowired
    private UserRepository userRepository;

    public User createUser(String name, String email) {
        User user = new User();
        user.setName(name);
        user.setEmail(email);
        user.setUsdBalance(new BigDecimal("1000"));
        user.setBtcBalance(BigDecimal.ZERO);
        return userRepository.save(user);
    }

    public void deleteUser(Long userId) {
        User user = userRepository.findById(userId)
                .orElseThrow(() -> new RuntimeException("User not found with id: " + userId));
        user.setDeleted(true);
        userRepository.save(user);
    }

    public User getUserById(Long userId) {
        return userRepository.findByIdAndDeletedFalse(userId)
                .orElseThrow(() -> new RuntimeException("User not found with id: " + userId));
    }

    public User updateUser(User user) {
        return userRepository.save(user);
    }
}